[refactor] #25 - 그룹 만들기 화면 뷰 분리
